ss: change type of CustomStateValue type ID to short
Matches with the org.eclipse.tracecompass change [1] below:

[1]https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203676

+++ callstack/org.eclipse.tracecompass.incubator.callstack.core/src/org/eclipse/tracecompass/incubator/callstack/core/base/EdgeStateValue.java
@@ -35,7 +35,7 @@ public class EdgeStateValue extends CustomStateValue {
     /**
      * The Custom Type value ID for {@link EdgeStateValue}s.
      */
-    public static final byte CUSTOM_TYPE_ID = 22;
+    public static final short CUSTOM_TYPE_ID = 22;
     /**
      * The {@link CustomStateValueFactory} for {@link EdgeStateValue}s
      */
@@ -113,7 +113,7 @@ public class EdgeStateValue extends CustomStateValue {
     }
 
     @Override
-    protected @NonNull Byte getCustomTypeId() {
+    protected @NonNull Short getCustomTypeId() {
         return CUSTOM_TYPE_ID;
     }
