+++ tracetypes/org.eclipse.tracecompass.incubator.ros2.core/src/org/eclipse/tracecompass/incubator/internal/ros2/core/model/HostProcessPointer.java
@@ -27,7 +27,7 @@ import org.eclipse.tracecompass.datastore.core.serialization.ISafeByteBufferWrit
 public class HostProcessPointer extends HostProcessValue<@NonNull Long> {
 
     /** Custom type value ID for this object */
-    public static final byte CUSTOM_TYPE_ID_HOST_PROCESS_POINTER = 73;
+    public static final short CUSTOM_TYPE_ID_HOST_PROCESS_POINTER = 73;
     /** CustomStateValueFactory for this object */
     @SuppressWarnings("restriction")
     public static final @NonNull CustomStateValueFactory HOST_PROCESS_POINTER_VALUE_FACTORY = b -> HostProcessPointer.read(b);
@@ -78,7 +78,7 @@ public class HostProcessPointer extends HostProcessValue<@NonNull Long> {
     }
 
     @Override
-    protected @NonNull Byte getCustomTypeId() {
+    protected @NonNull Short getCustomTypeId() {
         return CUSTOM_TYPE_ID_HOST_PROCESS_POINTER;
     }
