+++ tracetypes/org.eclipse.tracecompass.incubator.ros2.core/src/org/eclipse/tracecompass/incubator/internal/ros2/core/model/messages/Ros2PubInstance.java
@@ -27,7 +27,7 @@ import com.google.common.base.Objects;
 public class Ros2PubInstance extends Ros2Instance<@NonNull Ros2ObjectHandle> {
 
     /** Custom type value ID for this object */
-    public static final byte CUSTOM_TYPE_ID = 65;
+    public static final short CUSTOM_TYPE_ID = 65;
     /** CustomStateValueFactory for this object */
     @SuppressWarnings("restriction")
     public static final @NonNull CustomStateValueFactory ROS2_PUB_INSTANCE_VALUE_FACTORY = Ros2PubInstance::read;
@@ -112,7 +112,7 @@ public class Ros2PubInstance extends Ros2Instance<@NonNull Ros2ObjectHandle> {
     }
 
     @Override
-    protected @NonNull Byte getCustomTypeId() {
+    protected @NonNull Short getCustomTypeId() {
         return CUSTOM_TYPE_ID;
     }
