+++ tracetypes/org.eclipse.tracecompass.incubator.ros2.core/src/org/eclipse/tracecompass/incubator/internal/ros2/core/model/objects/Ros2CallbackObject.java
@@ -27,7 +27,7 @@ import com.google.common.base.Objects;
 public class Ros2CallbackObject extends Ros2Object<@NonNull HostProcessPointer> {
 
     /** Custom type value ID for this object */
-    public static final byte CUSTOM_TYPE_ID = 64;
+    public static final short CUSTOM_TYPE_ID = 64;
     /** CustomStateValueFactory for this object */
     @SuppressWarnings("restriction")
     public static final @NonNull CustomStateValueFactory ROS2_CALLBACK_OBJECT_VALUE_FACTORY = b -> Ros2CallbackObject.read(b);
@@ -122,7 +122,7 @@ public class Ros2CallbackObject extends Ros2Object<@NonNull HostProcessPointer> 
     }
 
     @Override
-    protected @NonNull Byte getCustomTypeId() {
+    protected @NonNull Short getCustomTypeId() {
         return CUSTOM_TYPE_ID;
     }
