+++ tracetypes/org.eclipse.tracecompass.incubator.ros2.core/src/org/eclipse/tracecompass/incubator/internal/ros2/core/model/objects/Ros2PublisherObject.java
@@ -23,7 +23,7 @@ import org.eclipse.tracecompass.incubator.internal.ros2.core.model.HostProcessPo
 public class Ros2PublisherObject extends Ros2PubSubObject {
 
     /** Custom type value ID for this object */
-    public static final byte CUSTOM_TYPE_ID = 61;
+    public static final short CUSTOM_TYPE_ID = 61;
     /** CustomStateValueFactory for this object */
     @SuppressWarnings("restriction")
     public static final @NonNull CustomStateValueFactory ROS2_PUBLISHER_OBJECT_VALUE_FACTORY = b -> Ros2PublisherObject.read(b);
@@ -61,7 +61,7 @@ public class Ros2PublisherObject extends Ros2PubSubObject {
     }
 
     @Override
-    protected @NonNull Byte getCustomTypeId() {
+    protected @NonNull Short getCustomTypeId() {
         return CUSTOM_TYPE_ID;
     }
